add wrap config option, default to false
Default to not wrapping long function names etc that are too big for the window.

If wrap=true is set, use some sane linebreak/breakindent/showbreak
options to wrap with an indent at the current indent level of the UI
tree, instead of just wrapping back to the far left side of the window.